Refactor project structure: move files from /utils to /helpers
