+++ src/lib/helpers/calculateCheckDigit.ts
